+++ sdk/resourcemanager/resources/armresources/client.go
@@ -46,7 +46,7 @@ func NewClient(subscriptionID string, credential azcore.TokenCredential, options
 
 // CheckExistence - Checks whether a resource exists.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The name of the resource group containing the resource to check. The name is case insensitive.
 //   - resourceProviderNamespace - The resource provider of the resource to check.
 //   - parentResourcePath - The parent resource identity.
@@ -107,9 +107,11 @@ func (client *Client) checkExistenceCreateRequest(ctx context.Context, resourceG
 	return req, nil
 }
 
-// CheckExistenceByID - Checks by ID whether a resource exists.
+// CheckExistenceByID - Checks by ID whether a resource exists. This API currently works only for a limited set of Resource
+// providers. In the event that a Resource provider does not implement this API, ARM will respond with
+// a 405. The alternative then is to use the GET API to check for the existence of the resource.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceID - The fully qualified ID of the resource, including the resource name and resource type. Use the format,
 //     /subscriptions/{guid}/resourceGroups/{resource-group-name}/{resource-provider-namespace}/{resource-type}/{resource-name}
 //   - apiVersion - The API version to use for the operation.
@@ -153,7 +155,7 @@ func (client *Client) checkExistenceByIDCreateRequest(ctx context.Context, resou
 // BeginCreateOrUpdate - Creates a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The name of the resource group for the resource. The name is case insensitive.
 //   - resourceProviderNamespace - The namespace of the resource provider.
 //   - parentResourcePath - The parent resource identity.
@@ -178,7 +180,7 @@ func (client *Client) BeginCreateOrUpdate(ctx context.Context, resourceGroupName
 // CreateOrUpdate - Creates a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) createOrUpdate(ctx context.Context, resourceGroupName string, resourceProviderNamespace string, parentResourcePath string, resourceType string, resourceName string, apiVersion string, parameters GenericResource, options *ClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginCreateOrUpdate"
@@ -238,7 +240,7 @@ func (client *Client) createOrUpdateCreateRequest(ctx context.Context, resourceG
 // BeginCreateOrUpdateByID - Create a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceID - The fully qualified ID of the resource, including the resource name and resource type. Use the format,
 //     /subscriptions/{guid}/resourceGroups/{resource-group-name}/{resource-provider-namespace}/{resource-type}/{resource-name}
 //   - apiVersion - The API version to use for the operation.
@@ -261,7 +263,7 @@ func (client *Client) BeginCreateOrUpdateByID(ctx context.Context, resourceID st
 // CreateOrUpdateByID - Create a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) createOrUpdateByID(ctx context.Context, resourceID string, apiVersion string, parameters GenericResource, options *ClientBeginCreateOrUpdateByIDOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginCreateOrUpdateByID"
@@ -304,7 +306,7 @@ func (client *Client) createOrUpdateByIDCreateRequest(ctx context.Context, resou
 // BeginDelete - Deletes a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The name of the resource group that contains the resource to delete. The name is case insensitive.
 //   - resourceProviderNamespace - The namespace of the resource provider.
 //   - parentResourcePath - The parent resource identity.
@@ -328,7 +330,7 @@ func (client *Client) BeginDelete(ctx context.Context, resourceGroupName string,
 // Delete - Deletes a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) deleteOperation(ctx context.Context, resourceGroupName string, resourceProviderNamespace string, parentResourcePath string, resourceType string, resourceName string, apiVersion string, options *ClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginDelete"
@@ -385,7 +387,7 @@ func (client *Client) deleteCreateRequest(ctx context.Context, resourceGroupName
 // BeginDeleteByID - Deletes a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceID - The fully qualified ID of the resource, including the resource name and resource type. Use the format,
 //     /subscriptions/{guid}/resourceGroups/{resource-group-name}/{resource-provider-namespace}/{resource-type}/{resource-name}
 //   - apiVersion - The API version to use for the operation.
@@ -406,7 +408,7 @@ func (client *Client) BeginDeleteByID(ctx context.Context, resourceID string, ap
 // DeleteByID - Deletes a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) deleteByID(ctx context.Context, resourceID string, apiVersion string, options *ClientBeginDeleteByIDOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginDeleteByID"
@@ -446,7 +448,7 @@ func (client *Client) deleteByIDCreateRequest(ctx context.Context, resourceID st
 // Get - Gets a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The name of the resource group containing the resource to get. The name is case insensitive.
 //   - resourceProviderNamespace - The namespace of the resource provider.
 //   - parentResourcePath - The parent resource identity.
@@ -520,7 +522,7 @@ func (client *Client) getHandleResponse(resp *http.Response) (ClientGetResponse,
 // GetByID - Gets a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceID - The fully qualified ID of the resource, including the resource name and resource type. Use the format,
 //     /subscriptions/{guid}/resourceGroups/{resource-group-name}/{resource-provider-namespace}/{resource-type}/{resource-name}
 //   - apiVersion - The API version to use for the operation.
@@ -573,7 +575,7 @@ func (client *Client) getByIDHandleResponse(resp *http.Response) (ClientGetByIDR
 
 // NewListPager - Get all the resources in a subscription.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - options - ClientListOptions contains the optional parameters for the Client.NewListPager method.
 func (client *Client) NewListPager(options *ClientListOptions) *runtime.Pager[ClientListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[ClientListResponse]{
@@ -626,7 +628,7 @@ func (client *Client) listCreateRequest(ctx context.Context, options *ClientList
 	if options != nil && options.Top != nil {
 		reqQP.Set("$top", strconv.FormatInt(int64(*options.Top), 10))
 	}
-	reqQP.Set("api-version", "2021-04-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -641,9 +643,85 @@ func (client *Client) listHandleResponse(resp *http.Response) (ClientListRespons
 	return result, nil
 }
 
+// NewListByParentPager - Lists child resources of a given parent resource.
+//
+// Generated from API version 2022-09-01
+//   - resourceGroupName - The name of the resource group containing the resource to get. The name is case insensitive.
+//   - resourceProviderNamespace - The namespace of the resource provider.
+//   - parentResourcePath - The parent resource identity.
+//   - resourceType - The resource type of the resource.
+//   - apiVersion - The API version to use for the operation.
+//   - options - ClientListByParentOptions contains the optional parameters for the Client.NewListByParentPager method.
+func (client *Client) NewListByParentPager(resourceGroupName string, resourceProviderNamespace string, parentResourcePath string, resourceType string, apiVersion string, options *ClientListByParentOptions) *runtime.Pager[ClientListByParentResponse] {
+	return runtime.NewPager(runtime.PagingHandler[ClientListByParentResponse]{
+		More: func(page ClientListByParentResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
+		},
+		Fetcher: func(ctx context.Context, page *ClientListByParentResponse) (ClientListByParentResponse, error) {
+			ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, "Client.NewListByParentPager")
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listByParentCreateRequest(ctx, resourceGroupName, resourceProviderNamespace, parentResourcePath, resourceType, apiVersion, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return ClientListByParentResponse{}, err
+			}
+			resp, err := client.internal.Pipeline().Do(req)
+			if err != nil {
+				return ClientListByParentResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return ClientListByParentResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listByParentHandleResponse(resp)
+		},
+		Tracer: client.internal.Tracer(),
+	})
+}
+
+// listByParentCreateRequest creates the ListByParent request.
+func (client *Client) listByParentCreateRequest(ctx context.Context, resourceGroupName string, resourceProviderNamespace string, parentResourcePath string, resourceType string, apiVersion string, options *ClientListByParentOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourcegroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{parentResourcePath}/{resourceType}"
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if resourceProviderNamespace == "" {
+		return nil, errors.New("parameter resourceProviderNamespace cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceProviderNamespace}", url.PathEscape(resourceProviderNamespace))
+	urlPath = strings.ReplaceAll(urlPath, "{parentResourcePath}", parentResourcePath)
+	urlPath = strings.ReplaceAll(urlPath, "{resourceType}", resourceType)
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", apiVersion)
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// listByParentHandleResponse handles the ListByParent response.
+func (client *Client) listByParentHandleResponse(resp *http.Response) (ClientListByParentResponse, error) {
+	result := ClientListByParentResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.ResourceListResult); err != nil {
+		return ClientListByParentResponse{}, err
+	}
+	return result, nil
+}
+
 // NewListByResourceGroupPager - Get all the resources for a resource group.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The resource group with the resources to get.
 //   - options - ClientListByResourceGroupOptions contains the optional parameters for the Client.NewListByResourceGroupPager
 //     method.
@@ -702,7 +780,7 @@ func (client *Client) listByResourceGroupCreateRequest(ctx context.Context, reso
 	if options != nil && options.Top != nil {
 		reqQP.Set("$top", strconv.FormatInt(int64(*options.Top), 10))
 	}
-	reqQP.Set("api-version", "2021-04-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -723,7 +801,7 @@ func (client *Client) listByResourceGroupHandleResponse(resp *http.Response) (Cl
 // on the groups until the move completes.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - sourceResourceGroupName - The name of the resource group from the source subscription containing the resources to be moved.
 //   - parameters - Parameters for moving resources.
 //   - options - ClientBeginMoveResourcesOptions contains the optional parameters for the Client.BeginMoveResources method.
@@ -746,7 +824,7 @@ func (client *Client) BeginMoveResources(ctx context.Context, sourceResourceGrou
 // on the groups until the move completes.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) moveResources(ctx context.Context, sourceResourceGroupName string, parameters MoveInfo, options *ClientBeginMoveResourcesOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginMoveResources"
@@ -784,7 +862,7 @@ func (client *Client) moveResourcesCreateRequest(ctx context.Context, sourceReso
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-04-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -796,7 +874,7 @@ func (client *Client) moveResourcesCreateRequest(ctx context.Context, sourceReso
 // BeginUpdate - Updates a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceGroupName - The name of the resource group for the resource. The name is case insensitive.
 //   - resourceProviderNamespace - The namespace of the resource provider.
 //   - parentResourcePath - The parent resource identity.
@@ -821,7 +899,7 @@ func (client *Client) BeginUpdate(ctx context.Context, resourceGroupName string,
 // Update - Updates a resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) update(ctx context.Context, resourceGroupName string, resourceProviderNamespace string, parentResourcePath string, resourceType string, resourceName string, apiVersion string, parameters GenericResource, options *ClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginUpdate"
@@ -881,7 +959,7 @@ func (client *Client) updateCreateRequest(ctx context.Context, resourceGroupName
 // BeginUpdateByID - Updates a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - resourceID - The fully qualified ID of the resource, including the resource name and resource type. Use the format,
 //     /subscriptions/{guid}/resourceGroups/{resource-group-name}/{resource-provider-namespace}/{resource-type}/{resource-name}
 //   - apiVersion - The API version to use for the operation.
@@ -903,7 +981,7 @@ func (client *Client) BeginUpdateByID(ctx context.Context, resourceID string, ap
 // UpdateByID - Updates a resource by ID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) updateByID(ctx context.Context, resourceID string, apiVersion string, parameters GenericResource, options *ClientBeginUpdateByIDOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginUpdateByID"
@@ -950,7 +1028,7 @@ func (client *Client) updateByIDCreateRequest(ctx context.Context, resourceID st
 // error message. Retrieve the URL in the Location header value to check the result of the long-running operation.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 //   - sourceResourceGroupName - The name of the resource group from the source subscription containing the resources to be validated
 //     for move.
 //   - parameters - Parameters for moving resources.
@@ -976,7 +1054,7 @@ func (client *Client) BeginValidateMoveResources(ctx context.Context, sourceReso
 // error message. Retrieve the URL in the Location header value to check the result of the long-running operation.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-04-01
+// Generated from API version 2022-09-01
 func (client *Client) validateMoveResources(ctx context.Context, sourceResourceGroupName string, parameters MoveInfo, options *ClientBeginValidateMoveResourcesOptions) (*http.Response, error) {
 	var err error
 	const operationName = "Client.BeginValidateMoveResources"
@@ -1014,7 +1092,7 @@ func (client *Client) validateMoveResourcesCreateRequest(ctx context.Context, so
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-04-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {