+++ sdk/resourcemanager/resources/armresources/response_types.go
@@ -50,6 +50,11 @@ type ClientGetResponse struct {
 	GenericResource
 }
 
+// ClientListByParentResponse contains the response from method Client.NewListByParentPager.
+type ClientListByParentResponse struct {
+	ResourceListResult
+}
+
 // ClientListByResourceGroupResponse contains the response from method Client.NewListByResourceGroupPager.
 type ClientListByResourceGroupResponse struct {
 	ResourceListResult
@@ -446,7 +451,7 @@ type ResourceGroupsClientUpdateResponse struct {
 	ResourceGroup
 }
 
-// TagsClientCreateOrUpdateAtScopeResponse contains the response from method TagsClient.CreateOrUpdateAtScope.
+// TagsClientCreateOrUpdateAtScopeResponse contains the response from method TagsClient.BeginCreateOrUpdateAtScope.
 type TagsClientCreateOrUpdateAtScopeResponse struct {
 	TagsResource
 }
@@ -461,7 +466,7 @@ type TagsClientCreateOrUpdateValueResponse struct {
 	TagValue
 }
 
-// TagsClientDeleteAtScopeResponse contains the response from method TagsClient.DeleteAtScope.
+// TagsClientDeleteAtScopeResponse contains the response from method TagsClient.BeginDeleteAtScope.
 type TagsClientDeleteAtScopeResponse struct {
 	// placeholder for future response values
 }
@@ -486,7 +491,7 @@ type TagsClientListResponse struct {
 	TagsListResult
 }
 
-// TagsClientUpdateAtScopeResponse contains the response from method TagsClient.UpdateAtScope.
+// TagsClientUpdateAtScopeResponse contains the response from method TagsClient.BeginUpdateAtScope.
 type TagsClientUpdateAtScopeResponse struct {
 	TagsResource
 }